+++ iotda/include/huaweicloud/iotda/v5/IoTDARegion.h
@@ -70,6 +70,10 @@ public:
             { 
                 "https://iotda.la-north-2.myhuaweicloud.com"
             });
+        regions["ru-moscow-1"] = Region("ru-moscow-1",
+            { 
+                "https://iotda.ru-moscow-1.myhuaweicloud.com"
+            });
         return regions;
     }
 private: