+++ iotdm/include/huaweicloud/iotdm/v5/IoTDMRegion.h
@@ -62,6 +62,10 @@ public:
             { 
                 "https://iotda-mgmt.la-south-2.myhuaweicloud.com"
             });
+        regions["ru-moscow-1"] = Region("ru-moscow-1",
+            { 
+                "https://iotda-mgmt.ru-moscow-1.myhuaweicloud.com"
+            });
         return regions;
     }
 private: